added projectDir flag, reporting used parameters and missing genes in the merged matrix
The user must define the project directory where the target file needs to be located, thus only target file name needs to be specified when running the script. The script also saves a txt file with any files that may be missing in the data dir and which are specified in the target file. It also generates a list of genes that may be missing in the merged matrix in case some genes are not present across all per-sample files